requirements.txt
pyyaml
PyGithub